final_touch

+++ push_swap/ft_create_stack.c
@@ -11,11 +11,11 @@
 /* ************************************************************************** */
 #include "push_swap.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	result;
+	long	i;
+	long	sign;
+	long	result;
 
 	sign = 1;
 	result = 0;
